requirements.txt
pandas
numpy
pillow
pypdf2
pdf2image
pytesseract
faiss-cpu
sentence-transformers
camelot-py
tabulate
openai
langchain
langchain-community
matplotlib